+++ angular_ast/CHANGELOG.md
@@ -1,5 +1,7 @@
 *   Added `ContainerAst`.
 
+*   Added support for annotations on `ContainerAst`.
+
 ## 0.5.2
 
 *   The `*` micro-syntax now supports newlines after an identifier.